extract app start into SR files
