+++ server/src/index.ts
@@ -1,56 +1,4 @@
 import "reflect-metadata";
 import "dotenv/config";
-import { createConnection } from "typeorm";
-import { ApolloServer } from "apollo-server-express";
-import * as express from "express";
-import * as session from "express-session";
-
-import { typeDefs } from "./typeDefs";
-import { resolvers } from "./resolvers";
-
-const startServer = async () => {
-  const server = new ApolloServer({
-    // These will be defined for both new or existing servers
-    typeDefs,
-    resolvers,
-    context: ({ req, res }: any) => ({ req, res })
-  });
-
-  let retries = 5;
-  while (retries) {
-    try {
-      await createConnection();
-      break;
-    } catch (err) {
-      console.log(err);
-      retries -= 1;
-      console.log(`retries left: ${retries}`);
-      // wait 5 seconds
-      await new Promise(res => setTimeout(res, 5000));
-    }
-  }
-
-  const app = express();
-
-  app.use(
-    session({
-      secret: "asdjlfkaasdfkjlads",
-      resave: false,
-      saveUninitialized: false
-    })
-  );
-
-  server.applyMiddleware({
-    app,
-    cors: {
-      credentials: true,
-      origin: "http://localhost:3000"
-    }
-  }); // app is from an existing express app
-
-  app.listen({ port: 4000 }, () =>
-    console.log(`🚀 Server ready at http://localhost:4000${server.graphqlPath}`)
-  );
-};
-
-startServer();
+import { start } from "./start";
+start();
