+++ server/src/start/app/session.ts
@@ -1,0 +1,10 @@
+import * as session from "express-session";
+export const createSession = (app: any) => {
+  app.use(
+    session({
+      secret: "asdjlfkaasdfkjlads",
+      resave: false,
+      saveUninitialized: false
+    })
+  );
+};
