+++ server/src/start/connect-db.ts
@@ -1,0 +1,15 @@
+import { createConnection } from "typeorm";
+export const connectDb = async (retries = 5) => {
+  while (retries) {
+    try {
+      await createConnection();
+      break;
+    } catch (err) {
+      console.log(err);
+      retries -= 1;
+      console.log(`retries left: ${retries}`);
+      // wait 5 seconds
+      await new Promise(res => setTimeout(res, 5000));
+    }
+  }
+};
