+++ server/src/start/index.ts
@@ -1,0 +1,7 @@
+import { connectDb } from "./connect-db";
+import { createApp } from "./app";
+
+export const start = async () => {
+  connectDb();
+  createApp();
+};
